\name{FDR2soilmoisture-package}
\alias{FDR2soilmoisture-package}
\alias{FDR2soilmoisture}
\docType{package}
\title{
\packageTitle{FDR2soilmoisture}
}
\description{
\packageDescription{FDR2soilmoisture}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{FDR2soilmoisture}
%\packageIndices{FDR2soilmoisture}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FDR2soilmoisture}

Maintainer: \packageMaintainer{FDR2soilmoisture}
}
%\references{
%~~ Literature or other references for background information ~~
%}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
